requirements.txt
torchvision ==0.1.6
torch==1.13.1



tqdm==4.64.1
aiofiles==23.1.0
alabaster==0.7.12
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.11.7
astropy==5.1
atomicwrites==1.4.0
attrs==22.1.0
Automat==20.2.0
autopep8==1.6.0
Babel==2.11.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
bitarray==2.5.1
bkcharts==0.2
black==22.6.0
bleach==4.1.0
bokeh==2.4.3
boto3==1.24.28
botocore==1.27.59
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.6
colorcet==3.0.1
comm==0.1.2
comtypes==1.1.10
constantly==15.1.0
contourpy==1.0.5
cookiecutter==1.7.3
cryptography==38.0.4
cssselect==1.1.0
cycler==0.11.0
Cython==0.29.32
cytoolz==0.12.0

entrypoints==0.4

fastjsonschema==2.16.2
filelock==3.9.0
flake8==4.0.1
Flask==2.2.2

h5py==3.7.0

joblib==1.1.1
json5==0.9.6
jsonschema==4.16.0
jupyter==1.0.0
jupyter_client==7.4.8
jupyter-console==6.4.4
jupyter_core==5.1.1
jupyter-server==1.23.4
jupyterlab==3.5.2
jupyterlab-pygments==0.1.2
jupyterlab_server==2.16.5
jupyterlab-widgets==1.0.0


matplotlib==3.6.2

nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2

numpy==1.24.1
numpydoc==1.5.0
olefile==0.46
opencv-python==4.7.0.72
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2

pickleshare==0.7.5
Pillow==9.3.0
pip==22.3.1
platformdirs==2.5.2
plotly==5.9.0
pluggy==1.0.0
ply==3.11

psutil==5.9.0
pyOpenSSL==22.0.0

requests==2.28.1
requests-file==1.5.1
sanic==22.12.0
Sanic-Cors==2.2.0
sanic-routing==22.8.0
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.9.3
Scrapy==2.6.2
seaborn==0.12.2

torch==1.13.1
torchaudio
torchvision

tqdm==4.64.1

wheel==0.37.1

